+++ cmd/deployIntegrationArtifact_test.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"bytes"
+	"fmt"
 	"io"
 	"io/ioutil"
 	"net/http"
@@ -66,8 +67,9 @@ func TestRunDeployIntegrationArtifact(t *testing.T) {
 			ResponseBody: ``, TestType: "Negative"}
 
 		err := runDeployIntegrationArtifact(&config, nil, &httpClient)
+		fmt.Printf("%s.\n", err)
 		// assert
-		assert.Error(t, err)
+		assert.EqualError(t, err, "Integration Flow deployment failed")
 	})
 
 }
