+++ pkg/cpi/CommonUtils.go
@@ -1,7 +1,6 @@
 package cpi
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -62,36 +61,4 @@ func (tokenParameters TokenParameters) GetBearerToken() (string, error) {
 	}
 	token := jsonResponse.Path("access_token").Data().(string)
 	return token, nil
-}
-
-//GetCPIFunctionMockResponse -Generate mock response payload for different CPI functions
-func GetCPIFunctionMockResponse(functionName, testType string) (*http.Response, error) {
-	switch functionName {
-	case "DeployIntegrationDesigntimeArtifact":
-		if testType == "Positive" {
-			res := http.Response{
-				StatusCode: 202,
-				Body:       ioutil.NopCloser(bytes.NewReader([]byte(``))),
-			}
-			return &res, nil
-		}
-		res := http.Response{
-			StatusCode: 500,
-			Body: ioutil.NopCloser(bytes.NewReader([]byte(`{
-						"code": "Internal Server Error",
-						"message": {
-						"@lang": "en",
-						"#text": "Cannot deploy artifact with Id 'flow1'!"
-						}
-					}`))),
-		}
-		return &res, errors.New("Internal Server Error")
-
-	default:
-		res := http.Response{
-			StatusCode: 404,
-			Body:       ioutil.NopCloser(bytes.NewReader([]byte(``))),
-		}
-		return &res, errors.New("Service not Found")
-	}
 }
